consolidate models
